refactoring vm_moduels for consistency
