+++ libs/vm_modules/include/vm_modules/vm_factory.hpp
@@ -26,9 +26,9 @@
 #include "vm_modules/math/random.hpp"
 
 #include "vm_modules/math/tensor.hpp"
-#include "vm_modules/ml/cross_entropy.hpp"
 #include "vm_modules/ml/graph.hpp"
-#include "vm_modules/ml/mean_square_error.hpp"
+#include "vm_modules/ml/ops/loss_functions/cross_entropy.hpp"
+#include "vm_modules/ml/ops/loss_functions/mean_square_error.hpp"
 
 namespace fetch {
 namespace vm_modules {
@@ -62,7 +62,7 @@ public:
     // ml modules - order is important!!
     math::CreateTensor(*module);
     ml::CreateGraph(*module);
-    ml::CreateCrossEntropy(*module);
+    fetch::vm_modules::ml::VMCrossEntropyLoss::Bind(*module);
     ml::CreateMeanSquareError(*module);
 
     return module;
